checkpoint of mess
